agentsociety2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentsociety2"
version = "2.1.2"
description = "A pure LLM-driven virtual world simulation platform (seed-driven world building, multi-agent simulation, intervention, reporting)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
keywords = [
    "agents",
    "llm",
    "simulation",
    "social-science",
    "multi-agent",
    "reinforcement-learning",
    "experiment",
    "artificial-intelligence",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.0.0",
    "numpy>=1.20.0,<2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.2",
    "json-repair>=0.46.2",
    "elemental-xenon>=1.1.0",
    "litellm>=1.82.0",
    "mcp[cli]>=1.13.1",
    "fastmcp>=2.12.2",
    "ruamel-yaml>=0.18.15",
    "prompt-toolkit>=3.0.51",
    "mem0ai>=1.0.5",
    "qdrant-client>=1.7.0",
    "aiohttp>=3.10.11",
    "stringcase>=1.2.0",
    "chromadb>=1.5.5",
    "fastapi>=0.128.0",
    "uvicorn>=0.35.0",
    "mineru[core]>=1.0.0",
    "torch>=2.9.1",
    "doclayout-yolo>=0.0.4",
    "ultralytics>=8.3.243",
    "transformers>=4.57.3",
    "ftfy>=6.3.1",
    "dill>=0.4.0",
    "pyclipper>=1.4.0",
    "black>=25.12.0",
    "sqlmodel>=0.0.16",
    "greenlet>=3.0.0",
    "pandas>=2.0.0",
    "ydata-profiling>=4.6.0",
    "sweetviz>=2.3.0",
    # Graph module dependencies
    "graphiti-core>=0.5.0",
    "neo4j>=5.20.0",
    "unstructured>=0.16.0",
]

[project.scripts]
agentsociety = "agentsociety2.society.cli:main"
agentsociety-workspace = "agentsociety2.society.workspace:main"

[project.urls]
Homepage = "https://github.com/tsinghua-fib-lab/agentsociety"
Documentation = "https://agentsociety2.readthedocs.io/"
Repository = "https://github.com/tsinghua-fib-lab/agentsociety.git"
Issues = "https://github.com/tsinghua-fib-lab/agentsociety/issues"
Changelog = "https://github.com/tsinghua-fib-lab/agentsociety/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.13",
    "mypy>=1.14.0",
    "pre-commit>=3.8.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-autobuild>=2024.9.19",
    "furo>=2024.8.6",
    "myst-parser>=4.0.0",
    "sphinx-autodoc2>=0.5.0",
    "sphinx-intl==2.1.0",
]
all = [
    "agentsociety2[dev,docs]",
]

[tool.hatch.build.targets.wheel]
packages = ["agentsociety2"]
# Explicitly include contrib env files that match test patterns
artifacts = [
  "agentsociety2/contrib/env/*_test.py",
  "agentsociety2/contrib/env/*_effect.py",
  "agentsociety2/contrib/env/*.py",
]

[tool.hatch.build.targets.sdist]
include = [
    "/agentsociety2",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]
artifacts = [
  "agentsociety2/contrib/env/*_test.py",
  "agentsociety2/contrib/env/*_effect.py",
  "agentsociety2/contrib/env/*.py",
]

[tool.hatch.build]
# Include all contrib env files
include = [
  "agentsociety2/contrib/env/*.py",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

# 注意：Mirothinker Agent 是可选功能，不在默认依赖中
# 如需使用，请参考 MiroThinker/README.md 进行配置
